\name{dsim-package}
\alias{dsim-package}
\alias{dsim}
\docType{package}
\title{dsim: a package for simulating a DNA computer}
\description{
The dsim package implements a rudimentary simulator of a DNA computer. The package is being developed and maintained in the free time of the authors. As a consequence, this package comes WITHOUT ANY WARRANTY regarding the consistency, completeness or correctness of the simulation and the software itself.
}
\details{
\tabular{ll}{
Package: \tab dsim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-20\cr
License: \tab GNU General Public License Version 3 (GPLv3)\cr
}
%% ~~ An overview of how to use the package, including the most important ~~
%% ~~ functions ~~
}
\author{
Author: Thorsten Tiede, Paul Duplys\cr
Maintainer: Thorsten Tiede <thorsten.tiede@gmail.com>, Paul Duplys <paul.duplys@gmail.com>\cr
}
\references{[1] Hoogeboom, ..., 1996
}

\keyword{ package 
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%%~~ the R documentation directory ~~
}
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
